add_jobs

+++ db/migrate/20220923104655_create_weathers.rb
@@ -1,10 +1,10 @@
 class CreateWeathers < ActiveRecord::Migration[6.1]
   def change
     create_table :weathers do |t|
-      t.integer :minimal_temperature
-      t.integer :maximal_temperature
-      t.integer :avg_temperature
-      t.integer :current_temperature
+      t.float :minimal_temperature
+      t.float :maximal_temperature
+      t.float :avg_temperature
+      t.float :current_temperature
       t.datetime :data
 
       t.timestamps
